refactor: renommage + nettoyage de fixme
