+++ server/tests/integration/common/actions/helpers/filters.test.ts
@@ -1,7 +1,5 @@
 import { strict as assert } from "assert";
 
-import { ObjectId } from "mongodb";
-
 import { buildMongoFilters, fullEffectifsFiltersConfigurations } from "@/common/actions/helpers/filters";
 
 const currentDate = new Date("2023-02-14T10:00:00Z");
